step5 피드백 반영
답변 삭제 시 controller에서는 id와 session부분만 AnswerService 로
넘겨주어 service로 위임

+++ src/main/java/com/codessquad/qna/service/AnswerService.java
@@ -35,9 +35,14 @@ public class AnswerService {
         return answerRepository.save(answer);
     }
 
-    public void delete(Answer answer) {
-        answer.delete();
-        answerRepository.save(answer);
+    public boolean delete(Long id, User loginUser) {
+        Answer answer = findById(id);
+        if (answer.isMatch(loginUser)) {
+            answer.delete();
+            answerRepository.save(answer);
+            return true;
+        }
+        return false;
     }
 
     public Answer findById(Long id) {
